Version 1.0.40 -Reworked tons of multithreading code (seriously... too many to list here)
