+++ About/About.xml
@@ -10,15 +10,8 @@
   <description>
     RimThreaded enables Rimworld to utilize multiple threads and thus greatly increases the speed of the game.
 
-    Version 1.0.39
-    -Added multithreading to FactionManager
-    -Added multithreading to SteadyEnvironmentEffects
-    -Added multithreading to WildPlantSpawner
-    -Added multithreading to WindManager
-    -Fixed bug in WorldPawns
-    -Fixed bug in WorldObjectsHolder
-    -Fixed bug in BiomeDef
-    -Fixed bug in Explosion AffectCell
+    Version 1.0.40
+    -Reworked tons of multithreading code (seriously... too many to list here)  
 
     DESCRIPTION:
     This project is still a work in progress and will likely not play well with other mods. I am uploading this hoping to get feedback and identify some bugs as I continue development. If you would like to contribute, I have provided a github link below. Submissions of bug reports with error logs are the most helpful to progress this project! Also submitting lists of known working and incompatible mods helps too.
