+++ Source/MaterialPool.cs
@@ -9,17 +9,17 @@ namespace RimThreaded
 		public static bool MatFrom(ref Material __result, MaterialRequest req)
 		{
 			Material material;
-			//if (TickList_Patch.materialResults.TryGetValue(req, out material))
+			//if (RimThreaded.materialResults.TryGetValue(req, out material))
 			//{
 			//__result = material;
 			//return false;
 			//}
 			int tID = Thread.CurrentThread.ManagedThreadId;
-			if (TickList_Patch.materialWaits.TryGetValue(tID, out EventWaitHandle eventWaitStart)) {
-				TickList_Patch.materialRequests.TryAdd(tID, req);
-				TickList_Patch.mainThreadWaitHandle.Set();
+			if (RimThreaded.materialWaits.TryGetValue(tID, out EventWaitHandle eventWaitStart)) {
+				RimThreaded.materialRequests.TryAdd(tID, req);
+				RimThreaded.mainThreadWaitHandle.Set();
 				eventWaitStart.WaitOne();
-				TickList_Patch.materialResults.TryGetValue(req, out material);
+				RimThreaded.materialResults.TryGetValue(req, out material);
 				__result = material;
 				return false;
 			}
