+++ Source/RimThreaded.csproj
@@ -73,6 +73,17 @@
   <ItemGroup>
     <Compile Include="AmbientSoundManager_Patch.cs" />
     <Compile Include="AttackTargetFinder.cs" />
+    <Compile Include="PawnRelationUtility.cs" />
+    <Compile Include="Map.cs" />
+    <Compile Include="TileTemperaturesComp.cs" />
+    <Compile Include="RimThreadedHarmony.cs" />
+    <Compile Include="RimThreadedMod.cs" />
+    <Compile Include="RimThreadedSettings.cs" />
+    <Compile Include="FilthMonitor2.cs" />
+    <Compile Include="WorldComponentUtility.cs" />
+    <Compile Include="DateNotifier.cs" />
+    <Compile Include="TradeShip.cs" />
+    <Compile Include="SeasonUtility.cs" />
     <Compile Include="BiomeDef.cs" />
     <Compile Include="WildPlantSpawner.cs" />
     <Compile Include="FactionManager.cs" />
