+++ Source/WorldObjectsHolder.cs
@@ -22,16 +22,8 @@ namespace RimThreaded
             AccessTools.FieldRefAccess<WorldObjectsHolder, List<WorldObject>>("worldObjects");
         public static bool WorldObjectsHolderTick(WorldObjectsHolder __instance)
         {
-            //tmpWorldObjects.Clear();
-            //tmpWorldObjects.AddRange(worldObjects(__instance));
-            //TickList_Patch.worldObjectsHolder = __instance;
-            TickList_Patch.worldObjects = worldObjects(__instance);
-            TickList_Patch.worldObjectsTicks = worldObjects(__instance).Count;
-            TickList_Patch.CreateMonitorThread();
-            TickList_Patch.monitorThreadWaitHandle.Set();
-            //TickList_Patch.MainThreadWaitLoop();
-            //for (int index = 0; index < tmpWorldObjects.Count; ++index)
-            //tmpWorldObjects[index].Tick();
+            RimThreaded.worldObjects = worldObjects(__instance);
+            RimThreaded.worldObjectsTicks = worldObjects(__instance).Count;
             return false;
         }
